WBO/WBO-collection/requirements.txt
matplotlib~=3.7.1
numpy~=1.24.3
biopython
hmmlearn~=0.3.0
requests